Fixed data types from int to Long

+++ src/main/java/com/ftms/ftmsapi/controller/TaskController.java
@@ -17,7 +17,7 @@ public class TaskController {
     TaskRepository taskRepository;
 
     @RequestMapping("/tasks/get")
-    public List<Task> getTasksByJobEmployee(@Valid @RequestBody int job_id, int employee_id) {
+    public List<Task> getTasksByJobEmployee(@Valid @RequestBody Long job_id, Long employee_id) {
         return taskRepository.getByJobAndEmployee(job_id, employee_id);
     }
 }
